+++ info.rkt
@@ -1,7 +1,7 @@
 #lang info
 (define collection "skia")
-(define version "0.10.1")
-(define pkg-desc "Experimental standalone CPU Skia drawing, HarfBuzz shaping, font management/text blobs, expanded paths/SVG geometry, filters, pictures/recording, vector effects, text, shaders, and image codecs through the SkiaSharp C ABI")
+(define version "0.11.0")
+(define pkg-desc "Experimental standalone CPU Skia drawing, paragraph text layout, HarfBuzz shaping, font management/text blobs, expanded paths/SVG geometry, filters, pictures/recording, vector effects, text, shaders, and image codecs through the SkiaSharp C ABI")
 (define deps '(("base" #:version "8.7") "draw-lib"))
 (define build-deps '("rackunit-lib"))
 (define license 'MIT)